+++ R/sanity.R
@@ -94,6 +94,19 @@ sanitize_output <- function(output) {
     out <- output
   }
 
+  if (isTRUE(check_dependency("litedown"))) {
+    fmt <- tryCatch(litedown::get_context("format"), error = function(e) NULL)
+    if (identical(fmt, "latex")) {
+      return("latex")
+    } else if (identical(fmt, "markdown")) {
+      return("markdown") 
+    } else if (identical(fmt, "commonmark")) {
+      return("markdown")
+    } else if (identical(fmt, "html")) {
+      return("html")
+    }
+  }
+
   if (isTRUE(check_dependency("knitr"))) {
     if (isTRUE(knitr::pandoc_to() %in% c("latex", "beamer"))) {
       flag <- getOption("tinytable_latex_preamble", default = TRUE)